Finished implementing click basic functionality, for now I need to add better user feedback and --help for the command
