--- __init__.py
@@ -1,10 +1,0 @@
-from .formatter import DockingFormatter
-
-"""
-DockingFormatter
-
-An simple formatter for docking logs.
-"""
-
-__version__ = "0.1.0"
-__author__ = 'Łukasz Stachnik'
